Inventory system completely reworked. Quests dropped.

+++ src/L2dotNET.GameService/model/items/ItemHandler.cs
@@ -94,7 +94,7 @@ namespace L2dotNET.GameService.Model.Items
                         {
                             string str = m.Attribute("exchange").Value;
                             foreach (string st in str.Split(';'))
-                                ih.addExchangeItem(Convert.ToInt32(st.Split('-')[0]), Convert.ToInt64(st.Split('-')[1]));
+                                ih.addExchangeItem(Convert.ToInt32(st.Split('-')[0]), Convert.ToInt32(st.Split('-')[1]));
                         }
 
                         if (m.Attribute("pet") != null)
