+++ src/L2dotNET.GameService/tables/admin_bypass/ab_teleport_entry.cs
@@ -7,7 +7,7 @@ namespace L2dotNET.GameService.Tables.Admin_Bypass
         public int y;
         public int z;
         public int id;
-        public long cost = 0;
+        public int cost = 0;
         public int itemId = 57;
     }
 }
