+++ src/L2dotNET.GameService/tables/recipe_item_entry.cs
@@ -5,10 +5,10 @@ namespace L2dotNET.GameService.Tables
     public class recipe_item_entry
     {
         public ItemTemplate item;
-        public long count;
+        public int count;
         public double rate;
 
-        public recipe_item_entry(int id, long count)
+        public recipe_item_entry(int id, int count)
         {
             item = ItemTable.Instance.GetItem(id);
             this.count = count;
